toolkit/content/tests/chrome/test_labelcontrol.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  XUL Widget Test for label control="value"
  -->
<window title="tabindex" width="500" height="600"
        onload="runTests()"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>  

<label id="lab" control="ctl"/>
<textbox id="ctl" value="Test"/>
<checkbox id="chk" value="Checkbox"/>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

<script>
<![CDATA[

SimpleTest.waitForExplicitFinish();
 
function runTests()
{
  is($("lab").control, "ctl", "control");
  is($("lab").labeledControlElement, $("ctl"), "labeledControlElement");
  is($("ctl").labelElement, $("lab"), "labelElement");
  is($("chk").labelElement.className, "checkbox-label", "labelElement");

  SimpleTest.finish();
}

]]>

</script>

</window>
